+++ packages/linode-js-sdk/src/kubernetes/types.ts
@@ -1,10 +1,12 @@
 export interface KubernetesCluster {
   created: string;
+  updated: string;
   region: string;
   status: string; // @todo enum this
   label: string;
-  version: string;
+  k8s_version: string;
   id: number;
+  tags: string[];
 }
 
 export interface KubeNodePoolResponse {
@@ -34,12 +36,12 @@ export interface KubernetesVersion {
 }
 
 export interface KubernetesEndpointResponse {
-  endpoints: string[];
+  endpoint: string;
 }
 
 export interface CreateKubeClusterPayload {
   label?: string; // Label will be assigned by the API if not provided
   region?: string; // Will be caught by Yup if undefined
   node_pools: PoolNodeRequest[];
-  version?: string; // Will be caught by Yup if undefined
+  k8s_version?: string; // Will be caught by Yup if undefined
 }