+++ packages/manager/src/components/EnhancedNumberInput/EnhancedNumberInput.tsx
@@ -49,6 +49,9 @@ const useStyles = makeStyles((theme: Theme) => ({
       height: 30,
       minWidth: 30
     },
+    '& $input': {
+      padding: '0 8px'
+    },
     '& $textField': {
       width: 50,
       minWidth: 40,
@@ -67,25 +70,45 @@ interface Props {
   value: number;
   setValue: (value: number) => void;
   disabled?: boolean;
+  max?: number;
+  min?: number;
 }
 
 type FinalProps = Props;
 
 export const EnhancedNumberInput: React.FC<FinalProps> = props => {
-  const { inputLabel, small, value, setValue, disabled } = props;
+  const { inputLabel, small, setValue, disabled } = props;
+
+  const max = props.max ?? 100;
+  const min = props.min ?? 0;
+
+  let value = props.value;
+  if (value > max) {
+    value = max;
+  } else if (value < min) {
+    value = min;
+  }
 
   const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setValue(+e.target.value);
+    const parsedValue = +e.target.value;
+    if (parsedValue >= min && parsedValue <= max) {
+      setValue(+e.target.value);
+    }
   };
 
-  const incrementValue = () => setValue(value + 1);
+  const incrementValue = () => {
+    if (value < max) {
+      setValue(value + 1);
+    }
+  };
 
   const decrementValue = () => {
-    if (value > 0) {
+    if (value > min) {
       setValue(value - 1);
     }
   };
 
+  // TODO add error prop for error handling
   const classes = useStyles();
   return (
     <React.Fragment>
@@ -102,7 +125,7 @@ export const EnhancedNumberInput: React.FC<FinalProps> = props => {
           aria-label="Subtract 1"
           name="Subtract 1"
           onClick={decrementValue}
-          disabled={disabled || value === 0 ? true : false}
+          disabled={disabled || value === min}
           data-testid={'decrement-button'}
         >
           <Minus className={classes.minusIcon} />
@@ -121,9 +144,9 @@ export const EnhancedNumberInput: React.FC<FinalProps> = props => {
             className: classnames({
               [classes.input]: true
             }),
-            min: 0
+            min,
+            max
           }}
-          autoFocus={true}
           disabled={disabled}
           data-testid={'quantity-input'}
         />
@@ -134,7 +157,7 @@ export const EnhancedNumberInput: React.FC<FinalProps> = props => {
           aria-label="Add 1"
           name="Add 1"
           onClick={incrementValue}
-          disabled={disabled}
+          disabled={disabled || value === max}
           data-testid={'increment-button'}
         >
           <Plus className={classes.plusIcon} />
@@ -144,4 +167,4 @@ export const EnhancedNumberInput: React.FC<FinalProps> = props => {
   );
 };
 
-export default EnhancedNumberInput;
+export default React.memo(EnhancedNumberInput);